+++ client/src/App.jsx
@@ -5,8 +5,6 @@ import Home from "./pages/home";
 import "./App.css";
 
 function App() {
-  const [count, setCount] = useState(0);
-
   return (
     <>
       <Home />
