--- client/src/components/navbar.jsx
@@ -1,12 +1,0 @@
-const Navbar = () => {
-  return (
-    <div className="nav-bar-container">
-      <ul className="navbar">
-        <li>Projects</li>
-        <li>Skills</li>
-        <li>About Me</li>
-      </ul>
-    </div>
-  );
-};
-export default Navbar;
