+++ client/src/components/navbar/navbar.jsx
@@ -1,0 +1,11 @@
+
+const Navbar = () => {
+  return (
+    <ul className="navbar">
+      <li>Projects</li>
+      <li>Skills</li>
+      <li>About Me</li>
+    </ul>
+  );
+};
+export default Navbar;
