+++ client/src/components/navbar/style.jsx
@@ -1,0 +1,18 @@
+import styled from "styled-components";
+
+// export const Button = styled.button`
+//   /* Adapt the colors based on primary prop */
+//   background: ${(props) => (props.$primary ? "#BF4F74" : "white")};
+//   color: ${(props) => (props.$primary ? "white" : "#BF4F74")};
+
+//   font-size: 1em;
+//   margin: 1em;
+//   padding: 0.25em 1em;
+//   border: 2px solid #bf4f74;
+//   border-radius: 3px;
+// `;
+export const Container = styled.div`
+ margin: 0,
+ padding:0,
+ border:0
+`;
