+++ client/src/pages/style.jsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+  margin: 0;
+  border: 0;
+  padding: 0;
+  background: blue;
+`;
